backend-python/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0
pydantic[email]==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
bcrypt==4.1.3
python-multipart==0.0.9
celery==5.4.0
redis==5.0.4
httpx==0.27.0
python-docx==1.1.0
pypdf==4.2.0
sse-starlette==2.1.0
